requirements.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg2-binary==2.9.10
gunicorn==23.0.0
